Consolidating multiple test files into a single test that relies on user to set env variable
